moving & combining + removing check for if user created
